chore: move unnecessary components directly to the page
